requirements/common.txt
# Core Stuff
# -------------------------------------
Django==1.10.3
gunicorn==19.6.0

# Configuration
# -------------------------------------
django-environ==0.4.1
python-dotenv==0.6.1

# Staticfiles
# -------------------------------------
whitenoise==3.2.2

# Extensions
# -------------------------------------
pytz==2016.7

# Models
# -------------------------------------
psycopg2==2.6.2
django-model-utils==2.6

Pillow
django-versatileimagefield==1.6.1
django-uuid-upload-path==1.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.5.3

# Authentication
# -------------------------------------
djangorestframework-jwt==1.8.0
django-axes==2.2.0

# django-rest-auth
django-rest-auth==0.8.2

# django-allauth
django-allauth==0.29.0

# Redis
# -------------------------------------
django-redis==4.6.0

# Celery
# -------------------------------------
celery==4.0.0
